Extract signal safe semaphore
